Darwix.ai/requirements.txt
Django>=5.0.6,<5.1.0
google-generativeai>=0.3.0,<0.4.0
python-dotenv>=1.0.0
transformers>=4.38.2
torch>=2.2.0
pyannote.audio>=3.3.2
numpy>=1.25.0
soundfile>=0.12.1
pydub>=0.25.1
huggingface-hub>=0.20.0
django-cors-headers>=4.3.0
requests>=2.31.0